fix: circular reference
