+++ tns-core-modules/application/application-common.ts
@@ -40,7 +40,9 @@ import {
     LoadAppCSSEventData,
     UnhandledErrorEventData
 } from "./application";
+
 import { DeviceOrientation } from "../ui/enums/enums";
+import { CSS_CLASS_PREFIX } from "../ui/utils";
 
 export { UnhandledErrorEventData, DiscardedErrorEventData, CssChangedEventData, LoadAppCSSEventData };
 
@@ -53,13 +55,6 @@ export const lowMemoryEvent = "lowMemory";
 export const uncaughtErrorEvent = "uncaughtError";
 export const discardedErrorEvent = "discardedError";
 export const orientationChangedEvent = "orientationChanged";
-
-const MODAL = "modal";
-const ROOT = "root";
-
-export const CSS_CLASS_PREFIX = "ns-";
-export const MODAL_ROOT_VIEW_CSS_CLASS = `${CSS_CLASS_PREFIX}${MODAL}`;
-export const ROOT_VIEW_CSS_CLASSES = [`${CSS_CLASS_PREFIX}${ROOT}`];
 
 const ORIENTATION_CSS_CLASSES = [
     `${CSS_CLASS_PREFIX}${DeviceOrientation.portrait}`,
