+++ tns-core-modules/ui/utils-common.ts
@@ -1,0 +1,6 @@
+﻿const MODAL = "modal";
+const ROOT = "root";
+
+export const CSS_CLASS_PREFIX = "ns-";
+export const MODAL_ROOT_VIEW_CSS_CLASS = `${CSS_CLASS_PREFIX}${MODAL}`;
+export const ROOT_VIEW_CSS_CLASSES = [`${CSS_CLASS_PREFIX}${ROOT}`];
